+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Enhancements
 
+* Added `static RealmObject.getRealm(RealmModel)`, `RealmObject.getRealm() and `DynamicRealmObject.getDynamicRealm()` (#4720).
+
 ### Bug Fixes
 
 ### Internal
@@ -72,7 +74,6 @@
 * Added support for querying inverse relationships (#2904).
 * Moved inverse relationships out of beta stage.
 * Added `Realm.getDefaultConfiguration()` (#4725).
-* Added `RealmObject.getConfiguration()` and `static RealmObject.getConfiguration(RealmModel)` (#4720).
 
 ### Bug Fixes